// src/app/shared/generated/model/aIKnowledgeVectorDb.ts
/**
 * onecx-ai-ui-bff
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AIContext } from './aIContext';


export interface AIKnowledgeVectorDb { 
    modificationCount?: number;
    id: string;
    name?: string;
    description?: string;
    vdb?: string;
    vdbCollection?: string;
    aiContext?: AIContext;
}